requirements.txt
streamlit
python-docx
pandas
openai>=1.0.0 